Freely join and leave in character selection menu. Leaving automatically reorganizes the player list and IDs. Limited players to having unique player models(i.e. two players cannot be the warrior)
For testing purposes, max players is currently 2(ModelAbility.COUNT) because that is the current amount of active models. Will change this back to hard coded value of 4 when I have implemented other abilities and finalized animations for the current 2 models. That way I know what I need when importing and implementing other models.

Created a class for storing player information(PlayerInfo class)
-Stores input device, model, light ability, heavy ability, passive ability

Uses list of these objects to load in players appropriately